Add `GeocodeApi`
This requires the proper API key in credentials.

It provides a way to get a lat and lon from geocoding an address. We
just use a struct for easy results.

The main method we care abou `lat_lon_for` used `location`. Location
grabs the first search result to pull lat lon out of.