Add `CachedWeatherService`
This wraps `WeatherService` to add caching around the request.

The starting retention period is 3 days.

The key is the address searched for. Ideally, we'd like zip code, but
there are open questions. This'll get working while business needs are
sorted.

Along with the data, `fetched_at` is the unix seconds.  This will
allow us to do the freshness calculations later.